quant_engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart
celery[redis]==5.3.6
redis==5.0.1
httpx==0.26.0
numpy<2.0.0
pandas<2.0.0
scipy
scikit-learn
matplotlib
plotly
joblib
pyyaml
tqdm
loguru
fire
yfinance
pyqlib
